Refactor camera node to handle special case for Dabai Max
In the OBCameraNode class, the onNewFrameCallback method has been modified to handle a special case for the Dabai Max device. When the device's PID is DABAI_MAX_PID, the color extrinsic parameters are used instead of the depth extrinsic parameters. This change ensures that the correct intrinsic and distortion parameters are used for the Dabai Max device.